fix(issue-218): keep installer bridge bootstrap dependency-free
